requirements.txt
telethon
anthropic
python-dotenv
aiogram
